inngest function added for workspace

+++ backend/inngest/index.js
@@ -28,6 +28,12 @@ const syncUserDeletion = inngest.createFunction(
     {event: 'clerk/user.deleted'},
     async ({ event })=>{
         const {data} = event
+        const user = await prisma.user.findUnique({
+            where: { id: data.id, }
+        })
+        if (!user) {
+            return res.status(404).json({ message: "User not found" });
+        }
         await prisma.user.delete({
             where: {
                 id: data.id,
@@ -55,11 +61,93 @@ const syncUserUpdation = inngest.createFunction(
     }
 )
 
+// Inngest Function to save workspace data to a database
+const syncWorkspaceCreation = inngest.createFunction(
+    {id: 'sync-workspace-from-clerk' },
+    {event: 'clerk/organization.created'},
+    async ({ event }) => {
+        const { data } = event;
+        await prisma.workspace.create({
+            data: {
+                id: data.id,
+                name: data.name,
+                slug: data.slug,
+                ownerId: data.created_by,
+                image_url: data.image_url,
+            }
+        })
+        // Add creator as admin member
+        await prisma.workspaceMember.create({
+            data: {
+                userId: data.created_by,
+                workspaceId: data.id,
+                role: "ADMIN"
+            }
+        })
+    }
+)
+
+// Inggest Function to update workspace data in database
+const syncWorkspaceUpdation = inngest.createFunction(
+    { id: 'update-workspace-from-clerk' },
+    { event: 'clerk/organization.update' },
+    async ({ event }) => {
+        const { data } = event;
+        await prisma.workspace.update({
+            where: {
+                id: data.id
+            },
+            data: {
+                name: data.name,
+                slug: data.slug,
+                image_url: data.image_url,
+            }
+        })
+    }
+
+)
+
+// Inggest Function to delete workspace data in database
+const syncWorkspaceDeletion = inngest.createFunction(
+    { id: 'delete-workspace-with-clerk' },
+    { event: 'clerk/organization.deleted' },
+    async ({ event }) => {
+        const { data } = event;
+        await prisma.workspace.delete({
+            where: {
+                id: data.id
+            }
+        })
+    }
+)
+
+
+// Inggest Function to save workspace member data in database
+const syncWorkspaceMemberCreation = inngest.createFunction(
+    { id: 'sync-workspace-member-from-clerk' },
+    { event: 'clerk/organizationInvitation.accepted'},
+    async ({ event })=> {
+        const { data } = event;
+        await prisma.workspaceMember.create({
+            data: {
+                userId: data.user_id,
+                workspaceId: data.organization_id,
+                role: String(data.role_name).toUpperCase(),
+            }
+        })
+    }
+)
+
 
 // Create an empty array where we'll export future inngest functions
 export const functions = [
     syncUserCreation,
     syncUserDeletion,
-    syncUserUpdation
+    syncUserUpdation,
+    syncWorkspaceCreation,
+    syncWorkspaceUpdation,
+    syncWorkspaceDeletion,
+    syncWorkspaceMemberCreation
+
 ];
